+++ lbrynet/file_manager/EncryptedFileDownloader.py
@@ -27,28 +27,24 @@ def log_status(sd_hash, status):
         status_string = "finished"
     else:
         status_string = "unknown"
-    log.info("stream %s is %s", short_hash(sd_hash), status_string)
+    log.debug("stream %s is %s", short_hash(sd_hash), status_string)
 
 
 class ManagedEncryptedFileDownloader(EncryptedFileSaver):
     STATUS_RUNNING = "running"
     STATUS_STOPPED = "stopped"
     STATUS_FINISHED = "finished"
-    """
-    These are started by EncryptedFileManager, aka, file_manager
-    """
-    def __init__(self, rowid, stream_hash, peer_finder, rate_limiter,
-                 blob_manager, stream_info_manager, lbry_file_manager,
-                 payment_rate_manager, wallet, download_directory,
-                 file_name=None):
 
+    def __init__(self, rowid, stream_hash, peer_finder, rate_limiter, blob_manager,
+                 stream_info_manager, lbry_file_manager, payment_rate_manager, wallet,
+                 download_directory, sd_hash=None, key=None, stream_name=None,
+                 suggested_file_name=None):
         EncryptedFileSaver.__init__(self, stream_hash, peer_finder,
                                     rate_limiter, blob_manager,
                                     stream_info_manager,
                                     payment_rate_manager, wallet,
-                                    download_directory,
-                                    file_name)
-
+                                    download_directory, key, stream_name, suggested_file_name)
+        self.sd_hash = sd_hash
         self.rowid = rowid
         self.lbry_file_manager = lbry_file_manager
         self._saving_status = False
@@ -57,22 +53,16 @@ class ManagedEncryptedFileDownloader(EncryptedFileSaver):
     def saving_status(self):
         return self._saving_status
 
-    @defer.inlineCallbacks
-    def restore(self):
-
-        status = yield self.lbry_file_manager.get_lbry_file_status(self)
-        log_status(self.sd_hash, status)
-
+    def restore(self, status):
         if status == ManagedEncryptedFileDownloader.STATUS_RUNNING:
             # start returns self.finished_deferred
             # which fires when we've finished downloading the file
             # and we don't want to wait for the entire download
             self.start()
         elif status == ManagedEncryptedFileDownloader.STATUS_STOPPED:
-            defer.returnValue(False)
+            pass
         elif status == ManagedEncryptedFileDownloader.STATUS_FINISHED:
             self.completed = True
-            defer.returnValue(True)
         else:
             raise Exception("Unknown status for stream %s: %s" % (self.stream_hash, status))
 
@@ -147,8 +137,7 @@ class ManagedEncryptedFileDownloaderFactory(object):
         return True
 
     @defer.inlineCallbacks
-    def make_downloader(self, metadata, options, payment_rate_manager, download_directory=None,
-                        file_name=None):
+    def make_downloader(self, metadata, options, payment_rate_manager, download_directory=None):
         assert len(options) == 1
         data_rate = options[0]
         stream_hash = yield save_sd_info(self.lbry_file_manager.stream_info_manager,
@@ -156,9 +145,11 @@ class ManagedEncryptedFileDownloaderFactory(object):
         if metadata.metadata_source == StreamMetadata.FROM_BLOB:
             yield self.lbry_file_manager.save_sd_blob_hash_to_stream(stream_hash,
                                                                      metadata.source_blob_hash)
-        lbry_file = yield self.lbry_file_manager.add_lbry_file(stream_hash, payment_rate_manager,
+        lbry_file = yield self.lbry_file_manager.add_lbry_file(stream_hash,
+                                                               metadata.source_blob_hash,
+                                                               payment_rate_manager,
                                                                data_rate,
-                                                               download_directory, file_name)
+                                                               download_directory)
         defer.returnValue(lbry_file)
 
     @staticmethod